Prevent multiple validation from multiple events
Prior to this change, it was possible for the handler to process
multiple events in the same contextual run. This meant that if
a validator used fetch() there would be multiple fetch() for each event
triggered. This isn't ideal. Instead we ignore events while waiting on
the first validator to complete.

Code should be awaiting anyway so a race condition between different
values should not occur.