Allow event bubbling for validate
Prior to this change, we were bound to have the modifier be only used on
the input element as on other elements would remove it from the easy use
of validate().

But because events can bubble it is possible that wrapping elements can
respond to the event instead.

We also relied on the validated event to perform promise resolution.

This change fixes the awkward validate/validated event dance which
freed us from tying the event dispatch to only elements registered as
validatable and could be any element even those who capture a bubbled
event.

This will help in cases where some form components might not place the
...attributes on the actual input element.