Add watch= argument to deprecate validator-update
Prior to this change, we mixed a virtual event validator-update to help
track properties. This works currently because the validation workflow
allows ember to exercise the validator and consume tracked properties.

We will be doing a big refactoring which will break this assumption
requiring better and more explicit control on dependent properties.

This change introduces a new API and deprecates the old one. It will be
a step towards the 2.0.0 release.