Complete rewrite using vanilla validity lib
Prior to this change, the basic idea was implemented inline. Since then
a Vanilla version of these ideas have matured and we can leverage it to
lower the complexity of this addon. I've copied the vanilla
implementation into this code base and adjusted the modifiers/utilities
to match.

This also exposes the new verifyFormValidity utility as another
modifier.